+++ src/mympd_api/last_played.c
@@ -28,7 +28,8 @@
  */
 
 static sds get_last_played_obj(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        sds buffer, unsigned entity_count, int64_t last_played, const char *uri, struct t_list *expr_list, const struct t_fields *tagcols);
+        sds buffer, unsigned entity_count, int64_t last_played, const char *uri, struct t_list *expr_list,
+        const struct t_fields *tagcols, bool print_stickers);
 
 /**
  * Public functions
@@ -82,16 +83,15 @@ sds mympd_api_last_played_list(struct t_partition_state *partition_state, struct
 
     unsigned real_limit = offset + limit;
     struct t_list *expr_list = parse_search_expression_to_list(expression, SEARCH_TYPE_SONG);
-    if (partition_state->mpd_state->feat.stickers == true &&
-        tagcols->stickers.len > 0)
-    {
+    bool print_stickers = check_get_sticker(partition_state->mpd_state->feat.stickers, &tagcols->stickers);
+    if (print_stickers == true) {
         stickerdb_exit_idle(stickerdb);
     }
 
     struct t_list_node *current = partition_state->last_played.head;
     while (current != NULL) {
         obj = get_last_played_obj(partition_state, stickerdb, obj, entity_count, current->value_i,
-            current->key, expr_list, tagcols);
+            current->key, expr_list, tagcols, print_stickers);
         if (sdslen(obj) > 0) {
             if (entities_found >= offset) {
                 if (entities_returned++) {
@@ -109,9 +109,7 @@ sds mympd_api_last_played_list(struct t_partition_state *partition_state, struct
         current = current->next;
     }
     FREE_SDS(obj);
-    if (partition_state->mpd_state->feat.stickers == true &&
-        tagcols->stickers.len > 0)
-    {
+    if (print_stickers == true) {
         stickerdb_enter_idle(stickerdb);
     }
     free_search_expression_list(expr_list);
@@ -141,7 +139,8 @@ sds mympd_api_last_played_list(struct t_partition_state *partition_state, struct
  * @return pointer to buffer
  */
 static sds get_last_played_obj(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        sds buffer, unsigned entity_count, int64_t last_played, const char *uri, struct t_list *expr_list, const struct t_fields *tagcols)
+        sds buffer, unsigned entity_count, int64_t last_played, const char *uri, struct t_list *expr_list,
+        const struct t_fields *tagcols, bool print_stickers)
 {
     if (mpd_send_list_meta(partition_state->conn, uri)) {
         struct mpd_song *song;
@@ -151,9 +150,7 @@ static sds get_last_played_obj(struct t_partition_state *partition_state, struct
                 buffer = tojson_uint(buffer, "Pos", entity_count, true);
                 buffer = tojson_int64(buffer, "LastPlayed", last_played, true);
                 buffer = print_song_tags(buffer, partition_state->mpd_state, &tagcols->mpd_tags, song);
-                if (partition_state->mpd_state->feat.stickers == true &&
-                    tagcols->stickers.len > 0)
-                {
+                if (print_stickers == true) {
                     buffer = mympd_api_sticker_get_print_batch(buffer, stickerdb, STICKER_TYPE_SONG, mpd_song_get_uri(song), &tagcols->stickers);
                 }
                 buffer = sdscatlen(buffer, "}", 1);